refactor: extract dashboard-process.ts from dashboard-server.ts
Move PID file management and health probe functions into a dedicated
dashboard-process.ts module:
- getDashboardPidPath, writeDashboardServerInfo, readDashboardServerInfo,
  removeDashboardServerInfo, isDashboardServerRunning,
  findRunningDashboardServer, DashboardServerInfo type

dashboard-server.ts re-exports these for backward compatibility so
existing consumers (tests, MCP server) continue to work unchanged.
dashboard-lifecycle.ts now imports directly from dashboard-process.ts.

Reduces dashboard-server.ts from 645 to ~540 lines while creating a
focused, independently testable process management module.

Closes #601
