+++ packages/core/src/commands/dashboard-lifecycle.test.ts
@@ -24,7 +24,7 @@ const mockFindRunningDashboardServer = vi.fn();
 const mockIsDashboardServerRunning = vi.fn();
 const mockReadDashboardServerInfo = vi.fn();
 const mockRemoveDashboardServerInfo = vi.fn();
-vi.mock('./dashboard-server.js', () => ({
+vi.mock('./dashboard-process.js', () => ({
   findRunningDashboardServer: () => mockFindRunningDashboardServer(),
   isDashboardServerRunning: (port: number) => mockIsDashboardServerRunning(port),
   readDashboardServerInfo: () => mockReadDashboardServerInfo(),
